### 2022-10-09 #4829 @editorialbot
[PRE REVIEW]: Copper: a performance curve generator for building energy simulation
**Submitting author:** @lymereJ (<a href="http://orcid.org/0000-0002-1630-6886">Jeremy Lerond</a>)
**Repository:** https://github.com/pnnl/copper
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 0.1.5
