### 2022-10-11 #4829 @kthyng
Hi @lymereJ and thanks for your submission! I think @crvernon would be a good fit for editing this, but is also at PNNL. I know PNNL is a large organization so could you each state if you are connected in a more specific way than just both working there? We may be able to waive the COI after stating it, depending on the connection level.
