### 2022-10-12 #4829 @jasondegraw
@crvernon In principle yes, but @lymereJ and I are both on the [EnergyPlus](https://github.com/NREL/EnergyPlus) development team, so there's a potential COI there. I took a look at the COI policy and I'm not sure if it rises to that level, but thought I should probably bring it up. We haven't worked together directly, and while I have reviewed some work that he did a while back, I do believe that I can be impartial.
