### 2022-10-12 #4829 @samuelduchesne
> :wave: @samuelduchesne - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @crvernon I'd love to.
