### 2022-10-12 #4829 @crvernon
@jasondegraw thank you for your honest feedback.  Let me seek a different reviewer for this one since you are on the same development team though I do believe you could be impartial and appreciate your contributions in the past to JOSS.   Have a great day!
