### 2022-10-24 #4876 @editorialbot
[REVIEW]: Copper: a performance curve generator for building energy simulation
**Submitting author:** @lymereJ (<a href="http://orcid.org/0000-0002-1630-6886">Jeremy Lerond</a>)
**Repository:** https://github.com/pnnl/copper
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.1.6
