### 2022-12-01 #4876 @crvernon
📣 Mid-week rally!

👋 @samuelduchesne and @FWuellhorst could you provide an update to how things are going?  Also please let me know if you have any questions.

Keep up the good work!
