### 2022-12-03 pnnl/copper#53 @samuelduchesne
JOSS Review: Example JSON is malformed
This issue is linked to the JOSS Review: https://github.com/openjournals/joss-reviews/issues/4876.

In the *QuickStart Guide* Section of the documentation, the json example (in the command line section) is malformed:

```json
{
    "Quickstart_Guide_Chiller": {
    "eqp_type": "chiller",
    "compressor_type": "screw",
    "condenser_type": "water",
    "compressor_speed": "constant",
    "ref_cap": 300,
    "ref_cap_unit": "ton",
    "full_eff": 0.61,
    "full_eff_unit": "kw/ton",
    "part_eff": 0.52,
    "part_eff_unit": "kw/ton",
    "sim_engine": "energyplus",
    "model": "ect_lwt",
    "do": {
        "generate_set_of_curves": {
        "vars": ["eir-f-plr"],
        "method": "nearest_neighbor",
        "tol": 0.005,
        "export_path": "./",
        "export_format": "idf",
        "export_name": "Quickstart_Guide_Chiller"
        }
    }
}<<< missing a {
```
