### 2022-12-03 pnnl/copper#54 @samuelduchesne
JOSS Review: Command `copper run in.JSON` does not return what is expected from the documentation
This issue is linked to the JOSS Review: https://github.com/openjournals/joss-reviews/issues/4876.

Following the *QuickStart Guide* running on Windows, the command `copper run in.json` produces one file: `.idf`. It does not generate the curves information (json) as described in the documentation:

```json
# Taken from https://pnnl.github.io/copper/Quickstart%20Guide.html#using-copper-s-command-line-interface
{
    "Quickstart_Guide_Chiller": {
        "eir-f-t": {
        "type": "bi_quad",
        "ref_evap_fluid_flow": null,
        "ref_cond_fluid_flow": null,
        "ref_lwt": 6.67,
        "ref_ect": 29.4,
        "ref_lct": null,
        "units": "si",
        "x_min": 5.0,
        "y_min": 24.0,
        "x_max": 10.0,
        "y_max": 35.0,
        "out_min": 0.0,
        "out_max": null,
        "coeff1": 0.933884,
        "coeff2": -0.058212,
        "coeff3": 0.00450036,
        "coeff4": 0.00243,
        "coeff5": 0.000486,
        "coeff6": -0.001215,
        "coeff7": 0.0,
        "coeff8": 0.0,
        "coeff9": 0.0,
        "coeff10": 0.0
        },
        "cap-f-t": {
        "type": "bi_quad",
        "ref_evap_fluid_flow": null,
        "ref_cond_fluid_flow": null,
        "ref_lwt": 6.67,
        "ref_ect": 29.4,
        "ref_lct": null,
        "units": "si",
        "x_min": 5.0,
        "y_min": 24.0,
        "x_max": 10.0,
        "y_max": 35.0,
        "out_min": 0.0,
        "out_max": null,
        "coeff1": 0.257896,
        "coeff2": 0.0389016,
        "coeff3": -0.00021708,
        "coeff4": 0.0468684,
        "coeff5": -0.00094284,
        "coeff6": -0.00034344,
        "coeff7": 0.0,
        "coeff8": 0.0,
        "coeff9": 0.0,
        "coeff10": 0.0
        },
        "eir-f-plr": {
        "type": "quad",
        "ref_evap_fluid_flow": null,
        "ref_cond_fluid_flow": null,
        "ref_lwt": 6.67,
        "ref_ect": 29.4,
        "ref_lct": null,
        "units": "si",
        "x_min": 0.0,
        "y_min": 0.0,
        "x_max": 1.0,
        "y_max": 1.0,
        "out_min": 0.0,
        "out_max": null,
        "coeff1": 0.222903,
        "coeff2": 0.313387,
        "coeff3": 0.46371,
        "coeff4": 0.0,
        "coeff5": 0.0,
        "coeff6": 0.0,
        "coeff7": 0.0,
        "coeff8": 0.0,
        "coeff9": 0.0,
        "coeff10": 0.0
        }
    }
}
```

The `.idf` file produced is:

```
Curve:Biquadratic,
   _eir-f-t,
   0.4714919803545788,
   -0.0003413558515812992,
   -0.0006979572001416824,
   0.02997434988897861,
   -0.0002825958540954729,
   -0.000386000186145715,
   -999,
   999,
   -999,
   999,
   0,
   999;

Curve:Quadratic,
   _eir-f-plr,
   0.2307348541782999,
   0.1814196166337366,
   0.5873280781879636,
   -999,
   999,
   -999,
   999,
   0,
   999;

Curve:Biquadratic,
   _cap-f-t,
   0.9685131980885846,
   -0.004964252905445725,
   0.0009250222553367597,
   0.0005571266682425913,
   -8.731725051830714e-05,
   0.000421644919398898,
   -999,
   999,
   -999,
   999,
   0,
   999;

```
