### 2022-12-03 pnnl/copper#55 @samuelduchesne
JOSS Review: Adding `import copper as cp` at the top of each examples wouldn't hurt 
Many of the example code-blocks don't show the `import copper as cp`. The first code block at the top of each page should at least contain the import statement.
