### 2022-12-03 pnnl/copper#56 @samuelduchesne
JOSS Review: cp.chiller class does not have `random_seed` parameter
In the example detailed in https://pnnl.github.io/copper/Additional%20Examples.html#repeatability, a random seed is specified, but `random_seed` is not a parameter of the cp.chiller class.

```python
full_eff_target = 1.188
full_eff_target_alt = 1.178
part_eff_target = 0.876
part_eff_target_alt = 0.869

chlr = cp.chiller(
    compressor_type="scroll",
    condenser_type="water",
    compressor_speed="constant",
    ref_cap=100,
    ref_cap_unit="ton",
    full_eff=full_eff_target,
    full_eff_unit="kw/ton",
    full_eff_alt=full_eff_target_alt,
    full_eff_unit_alt="kw/ton",
    part_eff=part_eff_target,
    part_eff_unit="kw/ton",
    part_eff_ref_std="ahri_550/590",
    part_eff_alt=part_eff_target_alt,
    part_eff_unit_alt="kw/ton",
    part_eff_ref_std_alt="ahri_551/591",
    model="ect_lwt",
    sim_engine="energyplus",
    random_seed=1 <<< TypeError: __init__() got an unexpected keyword argument 'random_seed'
)
```
