### 2022-12-03 pnnl/copper#57 @samuelduchesne
JOSS Review: Capitalize class names is a convention
According to [PEP8](https://peps.python.org/pep-0008/#class-names), "Class names should normally use the CapWords convention.". This would prevent awkward references such as `copper.chiller.chiller`

![image](https://user-images.githubusercontent.com/22966009/205450620-24d0bbca-d814-43e4-bac9-ac8362a2a928.png)

Not a big deal, but worth mentioning...
