### 2022-12-03 pnnl/copper#58 @samuelduchesne
JOSS Review: Address citation format for reference in paper (Missing Author)
Can the first two references have an author (Main author or organization)? instead, the whole title is referenced as the author, which makes it hard to read.

![image](https://user-images.githubusercontent.com/22966009/205451763-dda67078-113a-4d0e-9f1d-8df90bd63dc9.png)
