### 2022-12-03 #4876 @samuelduchesne
Hi @lymereJ. First of all, congratulations on a quality submission that promises to be very helpful for energy modelers in academia and in the industry. My review focused on the python package and making sure that the documentation was consistent from the perspective of the user:

I encountered a couple issues along the way which need to be addressed:
1. https://github.com/pnnl/copper/issues/53
2. https://github.com/pnnl/copper/issues/54
3. https://github.com/pnnl/copper/issues/55
4. https://github.com/pnnl/copper/issues/56
5. https://github.com/pnnl/copper/issues/57

Regarding the paper, I believe the first 2 citations are missing an author (organization or main author); The title renders as the author which makes the whole thing awfully long:
- https://github.com/pnnl/copper/issues/58
