### 2022-12-04 pnnl/copper#59 @FWuellhorst
JOSS-Review: Use logging instead of prints
The code-base uses print-statements to inform the user. Depending on the application, this is not wanted:
> Make your code publish-ready
Let’s prepare your code for the uploading. First, you should remove all “print” statements from your code. It’s annoying when you work with a library and your command prompt is flooded with print messages that are not yours — therefore remove all of them. If you want to inform the user about certain activities, use [logging](https://docs.python.org/3/library/logging.html).

https://medium.com/@joel.barmettler/how-to-upload-your-python-package-to-pypi-65edc5fe9c56

Also, users would know what print statement is an error / info / debug -statement.
