### 2022-12-04 #4876 @FWuellhorst
Hi @lymereJ. During my review, I tried to understand the underlying methods used to calculate the performance curves. As I am not familiar with ASHRAE (more with European standards), I had a harder time understanding all abbreviations and required input data in the code. That being said, I have three main questions/concerns:
1. You state that copper is a tool which can be used for different software, but it seems to focus heavily on ASHRAE / EnergyPlus (standards and software from the states). For a broad international usage, it would be great to show the equations being solved somewhere, and document the meaning of certain options directly in the code-base (or at least refer to some ASHRAE document which explains the naming). Some small image or equation in the paper as well could help readers decide if copper fits their need or not.
2. Is the problem of underdetermined system of equations special to chillers? Or is your proposed method applicable to other devices as well, such as fans, air handling units or similar? Do you actually plan to incorporate such devices? If not, I would state that in the paper. Your work is nevertheless pretty impactful, even if it only focuses on chillers.
3. Did you check influence of GA settings on results? Documentations mentioned fixing the random seed. Out of interest, I ran the standard example 5 times and deviations only in the left part of `cap-f-t`. Showing such deviations when using the GA in one example would be good to teach users which GA-settings are important and which not.
![image](https://user-images.githubusercontent.com/22442514/205498527-38b51d03-a082-450c-a775-0990f8d8310f.png)

Regarding code-base: @samuelduchesne already mentioned some issues. I added another one regarding logging (https://github.com/pnnl/copper/issues/59). Besides that, the installation, examples and so forth work as documented and explain the usage (aside from my questions above) very well. Great work, thanks for your contribution!
