+++ copper/chiller.py
@@ -15,7 +15,7 @@ location = os.path.dirname(os.path.realpath(__file__))
 chiller_lib = os.path.join(location, "lib", "chiller_curves.json")
 
 
-class chiller:
+class Chiller:
     def __init__(
         self,
         ref_cap,
@@ -269,7 +269,7 @@ class chiller:
         num_nearest_neighbors=10,
         max_restart=None,
     ):
-        """Generate a set of curves for a particular chiller object.
+        """Generate a set of curves for a particular Chiller object.
 
         :param str method: Method used to generate the set of curves: `best_match`, `nearest_neighbor`, or `weighted_average`
         :param int pop_size: Population size used by the genetic algorithm
@@ -283,7 +283,7 @@ class chiller:
         :param float mutate: Probability of an individual to be mutated in the next generation
         :param tuple() bounds: Random modification bounds
         :param int random_seed: Integer that is used to inialize the random number generator, this should be used when identical results are needed from one run to another
-        :return: Set of curves object generated by the genetic algorithm that matches the chiller object definition
+        :return: Set of curves object generated by the genetic algorithm that matches the Chiller object definition
         :rtype: SetofCurves
 
         """
