### 2022-12-06 b4f98c2
Adjust doc config file.
