+++ .github/workflows/docs.yml
@@ -4,8 +4,8 @@ jobs:
   docs:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-python@v2
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v3
       - name: Update pip and install python dependencies
         run: |
           python -m pip install --upgrade pip
