### 2022-12-06 9d19699
Adjust doc config file.
