+++ .github/workflows/docs.yml
@@ -3,6 +3,10 @@ on: [push, pull_request, workflow_dispatch]
 jobs:
   docs:
     runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.9"]
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v3
