### 2022-12-06 5b23b82
Adjust doc config file.
