+++ .github/workflows/docs.yml
@@ -9,7 +9,7 @@ jobs:
         python-version: ["3.9"]
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-python@v3
+      - uses: actions/setup-python@v4
       - name: Update pip and install python dependencies
         run: |
           python -m pip install --upgrade pip
