### 2022-12-06 1e27229
Rename generator class as well.
