### 2022-12-06 3bf9f95
Move random_seed to the correct place.
