+++ docs/source/Additional Examples.rst
@@ -8,6 +8,8 @@ Most building energy modeling software use an entering condenser temperature (EC
 
 .. sourcecode:: python
 
+    import copper as cp
+
     chlr = cp.Chiller(
         compressor_type="scroll",
         condenser_type="water",
@@ -35,6 +37,8 @@ Targeting two different rating standards
 The rating conditions in AHRI Standards 550/590 and 551/591 are different. **Copper** supports the IPLV and performance curve generation for both standards. It is also possible to generate curves for two sets of targeted efficiencies, one for AHRI 550/590 and the other for 551/591 as demonstrated in the following example:
 
 .. sourcecode:: python
+
+    import copper as cp
 
     full_eff_target = 1.188
     full_eff_target_alt = 1.178
@@ -72,6 +76,8 @@ Repeatability
 Because **Copper** is used to find a solution to an underdetermined system of equations, there are a multitude (an infinite) number of solutions to a set of targeted equipment characteristics and efficiencies; hence, running **Copper** multiple times can lead to different sets of curves. For applications where repeatability is necessary, users can use the `random_seed` attribute in an equipment definition, which ensures that the same result is generated every time the same equipment definition is run using **Copper**.
 
 .. sourcecode:: python
+
+    import copper as cp
 
     chlr = cp.Chiller(
         ref_cap=300,
