+++ docs/paper.bib
@@ -1,5 +1,6 @@
 @standard {ahri_550590,
     title={550/590 Performance Rating of Water-chilling and Heat Pump Water-heating Packages Using the Vapor Compression Cycle},
+    author={AHRI},
     organization={AHRI},
     institution ={AHRI},
     address={Arlington, VA},
@@ -8,6 +9,7 @@
 
 @standard {ahri_551591,
     title={551/591 Performance Rating of Water-chilling and Heat Pump Water-heating Packages Using the Vapor Compression Cycle},
+    author={AHRI},
     organization={AHRI},
     institution ={AHRI},
     address={Arlington, VA},
@@ -16,6 +18,7 @@
 
 @standard {90.1,
     title={ANSI/ASHRAE/IES 90.1-2019, Energy Standard for Buildings Except Low-Rise Residential Buildings},
+    author={ASHRAE},
     organization={ASHRAE},
     institution={ASHRAE},
     address={Atlanta, GA},
