### 2022-12-07 1340ab6
added a setup to store common objects and teardown to clean up created files
