### 2022-12-07 pnnl/copper#58 @lymereJ
Hi @samuelduchesne - I believe that organization was already in. I added authors, but I still see the same outcome in the PDF. Here's the latest: https://github.com/pnnl/copper/actions/runs/3633532550. Do you know what could be the reason for that? I'll keep looking on my end as well.
