### 2022-12-08 ea869df
added some setup and teardown and removed unneeded import
