### 2022-12-08 0d1d728
started adding tests for unit class
