### 2022-12-09 b1df436
updated test class and added unit tests
