### 2022-12-09 2fb7f53
updated the constants and placed them in their own area for easier reference.
