+++ copper/constants.py
@@ -1,0 +1,13 @@
+"""
+constants.py
+====================================
+Holds all the constants referenced in copper.
+TODO: there are additional constants, but i haven't found them yet.
+"""
+KBTU_TO_KW = 3.412141633
+TON_TO_KBTU = 12.0
+KILO = 1000
+kEER = TON_TO_KBTU * KILO
+FIVE_OVER_NINE = 5/9
+NINE_OVER_FIVE = FIVE_OVER_NINE ** -1
+TEMP_CONSTANT = 32
