### 2022-12-09 4b8935c
started documenting the origin of the constants and left the origin units. it'd be too much to change all the unit calls right now.
