### 2022-12-09 533863c
minor readme change and made gitignore my local vscode stuff
