### 2022-12-10 125e7c4
started adding information about the input type
