### 2022-12-10 00053e2
filled in all i could about the type of input the CLI expects.
