### 2022-12-10 f11b013
replaced the print statements with logging
