+++ copper/chiller.py
@@ -11,6 +11,9 @@ from copper.units import Units, newUnits
 from copper.curves import *
 from copper.library import *
 import copper.constants as const
+from copper.constants import LOGGING_FORMAT
+import logging
+logging.basicConfig(format=LOGGING_FORMAT)
 
 location = os.path.dirname(os.path.realpath(__file__))
 chiller_lib = os.path.join(location, "lib", "chiller_curves.json")
@@ -496,7 +499,7 @@ class Chiller:
                     - Power: {round(kwpton * cap_op * cap_ton, 2)} kW,
                     - Efficiency: {round(kwpton, 3)} kW/ton
                     """
-                    print(part_report)
+                    logging.info(part_report)
 
                 # Store efficiency for IPLV calculation
                 kwpton_lst.append(eir / kbtu_to_kw * ton_to_kbtu)
@@ -516,7 +519,7 @@ class Chiller:
             )
 
             if output_report:
-                print(f"IPLV: {round(iplv, 3)} kW/ton")
+                logging.info(f"IPLV: {round(iplv, 3)} kW/ton")
         except:
             return -999
 
