### 2022-12-10 52b3b9c
replaced the print statements with logging
