+++ copper/units.py
@@ -79,41 +79,41 @@ class Units:
 
         """
         if new_unit == "kw/ton":
-            if self.unit == "cop":  # done
+            if self.unit == "cop":
                 return 12.0 / (self.value * 3.412)
-            elif self.unit == "kw/ton":  # done
+            elif self.unit == "kw/ton":
                 return self.value
-            elif self.unit == "eer":  # done
+            elif self.unit == "eer":
                 return 12.0 / self.value
             else:
                 return self.value
         elif new_unit == "cop":
-            if self.unit == "kw/ton":  # done
+            if self.unit == "kw/ton":
                 return 12.0 / self.value / 3.412
-            elif self.unit == "cop":  # done
+            elif self.unit == "cop":
                 return self.value
-            elif self.unit == "eer":  # done
+            elif self.unit == "eer":
                 return self.value / 3.412
             else:
                 return self.value
         elif new_unit == "eer":
-            if self.unit == "kw/ton":  # done
+            if self.unit == "kw/ton":
                 return 12.0 / self.value
-            elif self.unit == "eer":  # done
+            elif self.unit == "eer":
                 return self.value
-            elif self.unit == "cop":  # done
+            elif self.unit == "cop":
                 return 3.412 * self.value
             else:
                 return self.value
         elif new_unit == "ton":
-            if self.unit == "kW":  # done
+            if self.unit == "kW":
                 return self.value * (3412 / 12000)
-            elif self.unit == "W":  # done
+            elif self.unit == "W":
                 return self.value * (3.412 / 12000)
         elif new_unit == "kW":
-            if self.unit == "ton":  # done
+            if self.unit == "ton":
                 return self.value / (3412 / 12000)
-            if self.unit == "W":  # done
+            if self.unit == "W":
                 return self.value / 1000
         elif new_unit == "W":
             if self.unit == "ton":
