+++ copper/curves.py
@@ -14,8 +14,7 @@ import statsmodels.api as sm
 import statistics
 import itertools
 import logging
-
-from copper.units import Units
+from copper.units import *
 from copper.constants import LOGGING_FORMAT
 logging.basicConfig(format=LOGGING_FORMAT)
 
