### 2022-12-13 e0f0576
blacked the files and added some instruction set reader logic to the library file
