+++ tests/test_chiller.py
@@ -1,10 +1,12 @@
 from pathlib import Path
 from unittest import TestCase
 
-import copper as cp
+from copper.library import Library
+from copper.chiller import Chiller
+from copper.units import Units
 import pickle as pkl
 import numpy as np
-import CoolProp.CoolProp as CP
+from CoolProp import CoolProp
 import os
 
 LOCATION = os.path.dirname(os.path.realpath(__file__))
@@ -13,18 +15,21 @@ CHILLER_LIB = os.path.join(LOCATION, "../copper/lib", "chiller_curves.json")
 
 class TestChiller(TestCase):
     def setUp(self) -> None:
-        """Runs before every test. Good place to initialize values and store common objects.
-        """
-        self.lib = cp.Library(path=CHILLER_LIB)
+        """Runs before every test. Good place to initialize values and store common objects."""
+        self.lib = Library(path=CHILLER_LIB)
 
     def tearDown(self) -> None:
-        """Runs after every test and cleans up file created from the tests.
-        """
+        """Runs after every test and cleans up file created from the tests."""
         ...
+
+    def test_get_unique_eqp_fields(self):
+        unique_fields = self.lib.get_unique_curve_fields()
+        number_of_unique_min_plr = len(unique_fields["min_plr"])
+        assert number_of_unique_min_plr == 26
 
     def test_get_reference_variable(self):
 
-        chlr = cp.Chiller(
+        chlr = Chiller(
             compressor_type="centrifugal",
             condenser_type="water",
             compressor_speed="constant",
@@ -48,7 +53,7 @@ class TestChiller(TestCase):
             [34.6, 1.0] == [round(v, 1) for v in chlr.get_ref_values("eir-f-plr")]
         )
 
-        chlr = cp.Chiller(
+        chlr = Chiller(
             compressor_type="centrifugal",
             condenser_type="water",
             compressor_speed="constant",
@@ -72,7 +77,7 @@ class TestChiller(TestCase):
             [35.0, 1.0] == [round(v, 1) for v in chlr.get_ref_values("eir-f-plr")]
         )
 
-        chlr = cp.Chiller(
+        chlr = Chiller(
             compressor_type="centrifugal",
             condenser_type="water",
             compressor_speed="constant",
@@ -96,7 +101,7 @@ class TestChiller(TestCase):
             [1.0, 0.0] == [round(v, 1) for v in chlr.get_ref_values("eir-f-plr")]
         )
 
-        chlr = cp.Chiller(
+        chlr = Chiller(
             compressor_type="centrifugal",
             condenser_type="water",
             compressor_speed="constant",
@@ -120,7 +125,7 @@ class TestChiller(TestCase):
             [1.0, 0.0] == [round(v, 1) for v in chlr.get_ref_values("eir-f-plr")]
         )
 
-        chlr = cp.Chiller(
+        chlr = Chiller(
             compressor_type="centrifugal",
             condenser_type="air",
             compressor_speed="constant",
@@ -144,7 +149,7 @@ class TestChiller(TestCase):
             [1.0, 0.0] == [round(v, 1) for v in chlr.get_ref_values("eir-f-plr")]
         )
 
-        chlr = cp.Chiller(
+        chlr = Chiller(
             compressor_type="centrifugal",
             condenser_type="air",
             compressor_speed="constant",
@@ -172,7 +177,7 @@ class TestChiller(TestCase):
 
         curves = pkl.load(open("./tests/data/agg_curves.pkl", "rb"))
 
-        chlr = cp.Chiller(
+        chlr = Chiller(
             compressor_type="screw",
             condenser_type="water",
[…]
