### 2023-01-04 #4876 @lymereJ
@FWuellhorst - Sorry for the late reply and thanks again for the review, comments, and encouragements! Please, let me know if additional clarifications/efforts is needed.

> 1. You state that copper is a tool which can be used for different software, but it seems to focus heavily on ASHRAE / EnergyPlus (standards and software from the states). For a broad international usage, it would be great to show the equations being solved somewhere, and document the meaning of certain options directly in the code-base (or at least refer to some ASHRAE document which explains the naming). Some small image or equation in the paper as well could help readers decide if copper fits their need or not.

The tool currently determines curve coefficients for two chiller models which use polynomial curves: one that uses the chiller entering condenser fluid temperature (ECT) and another one that uses the chiller leaving condenser fluid temperature (LCT). Several available building energy simulation software other than EnergyPlus (e.g. TRNSYS, DOE-2, IES-VE) use one of the two chiller models, so the tool should be applicable to other energy modeling software as well.

While it currently only support the generation of performance curves for chillers, we're planning on expanding to other type of equipment so we've been hesitant to include chiller specific equations or references to the paper. That being said, we've modified the documentation to point to the EnergyPlus documentation about the chiller models so readers/potential users have a reference that document the equations behind the performance curves generated by the tool. See for example the links in the Quickstart Guide and the Additional Example sections.

> 2. Is the problem of underdetermined system of equations special to chillers? Or is your proposed method applicable to other devices as well, such as fans, air handling units or similar? Do you actually plan to incorporate such devices? If not, I would state that in the paper. Your work is nevertheless pretty impactful, even if it only focuses on chillers.

The problem of “undetermined system of equations” arise because the rating standards used in North America usually require some HVAC equipment to meet a full load efficiency and single part-load efficiency at rated conditions; and because how the way these values are calculated, there is are an infinite number of sets of curve coefficients that can satisfy these targets. While we have only demonstrated the results on chillers, this problem remains for other equipment as well. For example rooftop direct-expansion packages rated using the EER and IEER metrics have the same issue. The proposed methods used by the tool would generalize to these equipment.

Other types of equipment (that don't necessarily depend on specific full load and part-load energy efficiency metrics) could also benefit from the aggregation process offered by **Copper**. It enable users to generate *typical* performance curves for equipment by aggregating the performance of a few to a large number of similar equipment.

> 3. Did you check influence of GA settings on results? Documentations mentioned fixing the random seed. Out of interest, I ran the standard example 5 times and deviations only in the left part of `cap-f-t`. Showing such deviations when using the GA in one example would be good to teach users which GA-settings are important and which not.
>    ![image](https://user-images.githubusercontent.com/22442514/205498527-38b51d03-a082-450c-a775-0990f8d8310f.png)

Since there are no unique solutions (i.e. curve coefficients) that meet the full load efficiency and part-load efficiency requirements, there will be some variability with respect to the random seed. However, we can potentially reduce the variability to an extent by having a tighter (i.e. lower) tolerance. From a user perspective, a user could generate multiple sets of curve coefficients with respect to different random seeds. Subsequently, the user could pick a single set of curve coefficients from these sets based on a pre-defined criteria based on their needs (for instance, the user could pick the set that matches the closest to the target full load efficiency and part load efficiency).

Here's an example, 10 sets of curves generated for the same targets:

- `tol = 0.005`
![image](https://user-images.githubusercontent.com/5149279/210492075-6ac7529a-414e-439f-91a9-bcbb42925079.png)

- `tol = 0.1`
![image](https://user-images.githubusercontent.com/5149279/210492105-c4e6c1d7-b29a-483c-93d0-2968b31fb238.png)

> Regarding code-base: @samuelduchesne already mentioned some issues. I added another one regarding logging ([pnnl/copper#59](https://github.com/pnnl/copper/issues/59)). Besides that, the installation, examples and so forth work as documented and explain the usage (aside from my questions above) very well. Great work, thanks for your contribution!

In progress! Should be done before the end of the week.
