### 2023-01-04 #4876 @lymereJ
> @editorialbot generate pdf

@samuelduchesne  - I followed-up in https://github.com/pnnl/copper/issues/58 about the issue with the citations, I now have both authors and organization but the citations look similar. Do you know what could be the cause of it?

Thanks!
