### 2023-01-05 bc117f2
added basicconfig to init instead of in every file
