+++ copper/__init__.py
@@ -1,1 +1,5 @@
 from copper.chiller import *
+from copper.constants import LOGGING_FORMAT
+import sys
+
+logging.basicConfig(format=LOGGING_FORMAT, stream=sys.stdout)
