### 2023-01-05 11f0fd8
logging format no longer required
