+++ copper/curves.py
@@ -15,9 +15,6 @@ import statistics
 import itertools
 import logging
 from copper.units import *
-from copper.constants import LOGGING_FORMAT
-
-logging.basicConfig(format=LOGGING_FORMAT)
 
 
 class SetsofCurves:
