### 2023-01-05 beb35e1
added test for new logging
