### 2023-01-05 be56966
fixed logging test
