### 2023-01-05 c7f2ecd
Revert "added test for new logging"
This reverts commit beb35e1fc49b9243e1cdcb44d745f25dfb842509.
