### 2023-01-05 9607fc2
added test for logging
