+++ copper/constants.py
@@ -2,6 +2,5 @@
 constants.py
 ====================================
 Holds all the constants referenced in copper.
-TODO: there are additional constants, but i haven't found them yet.
 """
 LOGGING_FORMAT = "%(asctime)s - %(levelname)s - %(message)s"
