+++ copper/__init__.py
@@ -2,4 +2,4 @@ from copper.chiller import *
 from copper.constants import LOGGING_FORMAT
 import sys
 
-logging.basicConfig(format=LOGGING_FORMAT, stream=sys.stdout)
+logging.basicConfig(format=LOGGING_FORMAT, stream=sys.stdout, level=logging.INFO)
