+++ tests/test_logging.py
@@ -36,7 +36,7 @@ class TestLogging(TestCase):
             == "Target not met after 1 generations; Restarting the generator."
         )
         self.assertTrue(captured[0][0].levelname == "WARNING")
-        self.assertEquals(captured[0][1].msg, "GEN: 0, IPLV: 5.56, KW/TON: 5.2")
+        self.assertTrue("GEN: 0, IPLV: 5.56, KW/TON: 5" in captured[0][1].msg)
         self.assertTrue(captured[0][1].levelname == "INFO")
         self.assertTrue(
             captured[0][2].msg
