### 2023-01-06 8fad2ce
replaced the print statements with logging (issue #59) (#66)
* replaced print statements with logging
