+++ copper/constants.py
@@ -1,0 +1,6 @@
+"""
+constants.py
+====================================
+Holds all the constants referenced in copper.
+"""
+LOGGING_FORMAT = "%(asctime)s - %(levelname)s - %(message)s"
