### 2023-01-06 #4876 @lymereJ
> > Regarding code-base: @samuelduchesne already mentioned some issues. I added another one regarding logging ([pnnl/copper#59](https://github.com/pnnl/copper/issues/59)). Besides that, the installation, examples and so forth work as documented and explain the usage (aside from my questions above) very well. Great work, thanks for your contribution!
>
> In progress! Should be done before the end of the week.

@FWuellhorst - This has been implemented. Thanks for the suggestion!
