### 2023-01-29 #4876 @FWuellhorst
@crvernon Yes, sorry for the delay!

@lymereJ :
Thanks for the explanations and the updates in the documentation regarding the reference to the EP manual.

Overall, I am happy with this review as well!

My only concern remains regarding the scope of the paper itself.
You state that
> we've been hesitant to include chiller specific equations or references to the paper

, as you plan to extend for other devices, such as ` rooftop direct-expansion packages`. However, these devices are based on a refrigerant cycle as well.
The paper mentions HVAC and building energy simulation specifically. Thus, I expected to find performance curves for other typical HVAC devices as well, such as pumps, fans, or maybe devices like humidifiers etc.
Stating in the paper (and maybe the ReadMe) that you focus on AC (and possibly heating by means of heat pumps, should work as well) could even guide more users to your tool in the first place, as researchers may search explicitly for a device of interest.

Regarding the checklist, I marked everything complete, as this last bit is just naming wise, and nothing critical in my opinion.
