### 2023-01-29 #4876 @lymereJ
> My only concern remains regarding the scope of the paper itself. You state that
>
> > we've been hesitant to include chiller specific equations or references to the paper
>
> , as you plan to extend for other devices, such as ` rooftop direct-expansion packages`. However, these devices are based on a refrigerant cycle as well. The paper mentions HVAC and building energy simulation specifically. Thus, I expected to find performance curves for other typical HVAC devices as well, such as pumps, fans, or maybe devices like humidifiers etc. Stating in the paper (and maybe the ReadMe) that you focus on AC (and possibly heating by means of heat pumps, should work as well) could even guide more users to your tool in the first place, as researchers may search explicitly for a device of interest.

@FWuellhorst - Thanks for the note! I understand your concern. While I don't think we'll ever cover pumps or fans, we might end up covering other types of heating equipment such as for example boilers. Do you think that it would then be appropriate to edit the paper and README and mention that our focus is heating and cooling equipment?
