### 2023-01-30 #4876 @FWuellhorst
Regarding the reference: Maybe try using `@techreport` or just `misc` instead of `@standard`?
