+++ docs/paper.bib
@@ -1,4 +1,4 @@
-@standard {ahri_550590,
+@misc {ahri_550590,
     title={550/590 Performance Rating of Water-chilling and Heat Pump Water-heating Packages Using the Vapor Compression Cycle},
     author={AHRI},
     organization={AHRI},
@@ -7,7 +7,7 @@
     year={2020}
 }
 
-@standard {ahri_551591,
+@misc {ahri_551591,
     title={551/591 Performance Rating of Water-chilling and Heat Pump Water-heating Packages Using the Vapor Compression Cycle},
     author={AHRI},
     organization={AHRI},
@@ -16,7 +16,7 @@
     year={2020}
 }
 
-@standard {90.1,
+@misc {90.1,
     title={ANSI/ASHRAE/IES 90.1-2019, Energy Standard for Buildings Except Low-Rise Residential Buildings},
     author={ASHRAE},
     organization={ASHRAE},
