### 2023-02-01 #4876 @lymereJ
> Regarding the reference: Maybe try using `@techreport` or just `misc` instead of `@standard`?

That worked! It looks much better now. (@crvernon - FYI)
