### 2023-02-01 #4876 @lymereJ
> @lymereJ I think this would make things clearer for possible users!

I updated the paper, the doc, and the README. Thanks!
