### 2023-02-15 #4876 @crvernon
@lymereJ we are almost there!

Your paper is looking good!

So here is what we have left to do:

- [ ] Conduct a GitHub release of the current reviewed version of the software you now have on the main and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

- [ ] Please list the DOI of the archived version here
