### 2023-02-15 2f78516
Draft paper for JOSS submission (#51)
* Paper draft.

* Fix path to paper.

* Add missing commas.

* Test bid mod.

* Test bid mod.

* Add missing comma.

* Fix typo.

* Update bio.

* Add reference.

* Update paper and references.

* Misc changes.

* Fix typo.

* Editorial changes 1.

* Misc changes to refs.

* Misc changes to refs.

* Misc changes to refs.

* Add authors.

* Try misc instead of standarsd.

* Change HVAC to heating and cooling.

* Update Black formatting.
