+++ docs/paper.bib
@@ -1,0 +1,72 @@
+@misc {ahri_550590,
+    title={550/590 Performance Rating of Water-chilling and Heat Pump Water-heating Packages Using the Vapor Compression Cycle},
+    author={AHRI},
+    organization={AHRI},
+    institution ={AHRI},
+    address={Arlington, VA},
+    year={2020}
+}
+
+@misc {ahri_551591,
+    title={551/591 Performance Rating of Water-chilling and Heat Pump Water-heating Packages Using the Vapor Compression Cycle},
+    author={AHRI},
+    organization={AHRI},
+    institution ={AHRI},
+    address={Arlington, VA},
+    year={2020}
+}
+
+@misc {90.1,
+    title={ANSI/ASHRAE/IES 90.1-2019, Energy Standard for Buildings Except Low-Rise Residential Buildings},
+    author={ASHRAE},
+    organization={ASHRAE},
+    institution={ASHRAE},
+    address={Atlanta, GA},
+    year={2019}
+}
+
+@article{ashrae_conf_paper,
+    author={J Lerond},
+    title={Algorithmically Generated Chiller Performance Curves for Building Energy Simulation},
+    journal={ASHRAE transactions},
+    year={2019},
+    volume={125},
+    number={2},
+    pages={93}
+}
+
+@misc{comnet,
+    title={COMNET, Appendix H - Equipment Performance Curves},
+    author={COMNET},
+    howpublished={\url{https://comnet.org/}},
+    note={Accessed: 2022-09-22}
+}
+
+@misc{osti_1395882,
+    title={EnergyPlus™},
+    author={},
+    abstractNote={Originally developed in 1999. EnergyPlus™ is a whole building energy simulation program that engineers, architects, and researchers use to model both energy consumption—for heating, cooling, ventilation, lighting and plug and process loads—and water use in buildings. EnergyPlus is a console-based program that reads input and writes output to text files. It ships with a number of utilities including IDF-Editor for creating input files using a simple spreadsheet-like interface, EP-Launch for managing input and output files and performing batch simulations, and EP-Compare for graphically comparing the results of two or more simulations. Several comprehensive graphical interfaces for EnergyPlus are also available. DOE does most of its work with EnergyPlus using the OpenStudio® software development kit and suite of applications. DOE releases major updates to EnergyPlus twice annually.},
+    url={https://www.osti.gov//servlets/purl/1395882},
+    doi={},
+    url={https://www.osti.gov/biblio/1395882}
+}
+
+@software{doe2,
+  author={James J. Hirsch & Associates (JJH)},
+  title={DOE-2.2},
+  url={https://www.doe2.com/}
+}
+
+@misc{addendumbd,
+    author={ASHRAE},
+    title={Addendum bd to Standard 90.1-2019, Energy Standard for Buildings Except Low-Rise Residential Buildings},
+    institution={ASHRAE},
+    year={2022}
+}
+
+@misc{tspr,
+    title={TSPR Washington State Analysis Tool},
+    author={PNNL},
+    howpublished={\url{https://energycode.pnl.gov/HVACSystemPerformance/}},
+    note={Accessed: 2022-09-22}
+}
