+++ setup.py
@@ -6,7 +6,7 @@ this_directory = Path(__file__).parent
 long_description = (this_directory / "README.md").read_text()
 setup(
     name="copper-bem",
-    version="0.1.5",
+    version="0.1.6",
     description="Performance curve generator for building energy simulation",
     long_description=long_description,
     long_description_content_type="text/markdown",
