### 2023-02-17 #4876 @kyleniemeyer
Hi @lymereJ, I'm doing some final checks before publishing.

In the paper references, it looks like a number of your references use institutional/group authors, which are not currently being handled well. Any author name that doesn't follow the "First M. Last" format needs double `{{` and `}}` wrapped around it, otherwise BibTeX will try to interpret as a name. For example, `{James J. Hirsch & Associates (JJH)}` should be `{{James J. Hirsch & Associates}}` (no need to include the abbreviation inside the citation). The same should apply to any other organization/institution as an author.

Can you make those edits? Thanks!
