+++ docs/paper.bib
@@ -1,6 +1,6 @@
 @misc {ahri_550590,
     title={550/590 Performance Rating of Water-chilling and Heat Pump Water-heating Packages Using the Vapor Compression Cycle},
-    author={AHRI},
+    author={{AHRI}},
     organization={AHRI},
     institution ={AHRI},
     address={Arlington, VA},
@@ -9,7 +9,7 @@
 
 @misc {ahri_551591,
     title={551/591 Performance Rating of Water-chilling and Heat Pump Water-heating Packages Using the Vapor Compression Cycle},
-    author={AHRI},
+    author={{AHRI}},
     organization={AHRI},
     institution ={AHRI},
     address={Arlington, VA},
@@ -18,7 +18,7 @@
 
 @misc {90.1,
     title={ANSI/ASHRAE/IES 90.1-2019, Energy Standard for Buildings Except Low-Rise Residential Buildings},
-    author={ASHRAE},
+    author={{ASHRAE}},
     organization={ASHRAE},
     institution={ASHRAE},
     address={Atlanta, GA},
@@ -37,7 +37,7 @@
 
 @misc{comnet,
     title={COMNET, Appendix H - Equipment Performance Curves},
-    author={COMNET},
+    author={{COMNET}},
     howpublished={\url{https://comnet.org/}},
     note={Accessed: 2022-09-22}
 }
@@ -52,13 +52,13 @@
 }
 
 @software{doe2,
-  author={James J. Hirsch & Associates (JJH)},
+  author={{James J. Hirsch & Associates (JJH)}},
   title={DOE-2.2},
   url={https://www.doe2.com/}
 }
 
 @misc{addendumbd,
-    author={ASHRAE},
+    author={{ASHRAE}},
     title={Addendum bd to Standard 90.1-2019, Energy Standard for Buildings Except Low-Rise Residential Buildings},
     institution={ASHRAE},
     year={2022}
@@ -66,7 +66,7 @@
 
 @misc{tspr,
     title={TSPR Washington State Analysis Tool},
-    author={PNNL},
+    author={{Pacific Northwest National Laboratory}},
     howpublished={\url{https://energycode.pnl.gov/HVACSystemPerformance/}},
     note={Accessed: 2022-09-22}
 }
