### 2023-02-22 #4876 @lymereJ
> rences, it looks like a number of your references use institutional/group authors, which are not currently being handled well. Any aut

@kyleniemeyer - I just made the edits. Please, let me know if I need to address anything else. Thanks!
