### 2023-02-27 #4876 @kyleniemeyer
@lymereJ I was just about to ask about that—yes, `editorialbot` is using the version of the paper in the `joss_paper` branch, so that should contain the final/correct version of the paper
