### 2023-02-27 #4876 @kyleniemeyer
@openjournals/dev Hi, running into a weird error with publishing/accepting that appears to be a bug:
```
Run bundle exec ruby /home/runner/work/_actions/xuanxu/deposit-with-crossref-action/main/deposit_with_crossref.rb
[80](https://github.com/openjournals/joss-papers/actions/runs/4284571755/jobs/7461710124#step:7:86)
/home/runner/work/_actions/xuanxu/deposit-with-crossref-action/main/deposit_with_crossref.rb:15:in `<main>': undefined method `exists?' for File:Class (NoMethodError)
[81](https://github.com/openjournals/joss-papers/actions/runs/4284571755/jobs/7461710124#step:7:87)

[82](https://github.com/openjournals/joss-papers/actions/runs/4284571755/jobs/7461710124#step:7:88)
if File.exists?(crossref_filepath)
[83](https://github.com/openjournals/joss-papers/actions/runs/4284571755/jobs/7461710124#step:7:89)
       ^^^^^^^^
[84](https://github.com/openjournals/joss-papers/actions/runs/4284571755/jobs/7461710124#step:7:90)
Did you mean?  exist?
[85](https://github.com/openjournals/joss-papers/actions/runs/4284571755/jobs/7461710124#step:7:91)
Error: Process completed with exit code 1.
```
