### 2023-02-28 #4876 @xuanxu
Yep, clearly a bug. Working on it.
